refactor(eip): migrate ext command into product
