+++ cmd/ext_compat_test.go
@@ -6,21 +6,25 @@ import (
 	"testing"
 )
 
-func TestExtCommandDoesNotDependOnCompatShims(t *testing.T) {
-	src, err := os.ReadFile("ext.go")
+func TestExtCommandMigratedOutOfPlatformCmd(t *testing.T) {
+	if _, err := os.Stat("ext.go"); err == nil {
+		t.Fatal("cmd/ext.go must be removed after ext migrates to products/eip/internal/ext")
+	} else if !os.IsNotExist(err) {
+		t.Fatalf("stat ext.go: %v", err)
+	}
+
+	src, err := os.ReadFile("root.go")
 	if err != nil {
-		t.Fatalf("read ext.go: %v", err)
+		t.Fatalf("read root.go: %v", err)
 	}
-	for _, helper := range []string{
-		"describeUHostByID(",
-		"getUhostList(",
-		"sbindEIP(",
-		"unbindEIP(",
-	} {
-		if strings.Contains(string(src), helper) {
-			t.Fatalf("ext.go must not call compat helper %s", helper)
-		}
+	if contains := string(src); contains == "" {
+		t.Fatal("root.go is unexpectedly empty")
+	} else if strings.Contains(contains, "NewCmdExt(") {
+		t.Fatal("cmd/root.go must not register NewCmdExt after ext migrates to products/eip")
 	}
+}
+
+func TestExtCommandDoesNotDependOnCompatShims(t *testing.T) {
 	for _, path := range []string{"eip_compat.go", "uhost_compat.go"} {
 		if _, err := os.Stat(path); err == nil {
 			t.Fatalf("%s must be removed after ext owns its SDK helpers", path)
