+++ products/eip/internal/ext/status.go
@@ -1,0 +1,7 @@
+package ext
+
+const (
+	hostRunning = "Running"
+	hostStopped = "Stopped"
+	hostFail    = "Install Fail"
+)
