+++ products/eip/product.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/ucloud/ucloud-cli/pkg/cli"
 	internaleip "github.com/ucloud/ucloud-cli/products/eip/internal/eip"
+	internalext "github.com/ucloud/ucloud-cli/products/eip/internal/ext"
 )
 
 type product struct{}
@@ -12,9 +13,9 @@ type product struct{}
 func New() cli.Product { return product{} }
 
 func (product) Metadata() cli.Metadata {
-	return cli.Metadata{Name: "eip", Commands: []string{"eip"}}
+	return cli.Metadata{Name: "eip", Commands: []string{"eip", "ext"}}
 }
 
 func (product) NewCommand(ctx *cli.Context) []*cobra.Command {
-	return []*cobra.Command{internaleip.NewCommand(ctx)}
+	return []*cobra.Command{internaleip.NewCommand(ctx), internalext.NewCommand(ctx)}
 }
