+++ products/eip/testdata/completion.golden
@@ -22,3 +22,9 @@ ucloud eip release	region	dynamic
 ucloud eip unbind	eip-id	dynamic
 ucloud eip unbind	project-id	dynamic
 ucloud eip unbind	region	dynamic
+ucloud ext uhost switch-eip	create-eip-charge-type	static	Dynamic,Month,Trial,Year
+ucloud ext uhost switch-eip	create-eip-traffic-mode	static	Bandwidth,ShareBandwidth,Traffic
+ucloud ext uhost switch-eip	project-id	dynamic
+ucloud ext uhost switch-eip	region	dynamic
+ucloud ext uhost switch-eip	uhost-id	dynamic
+ucloud ext uhost switch-eip	zone	dynamic
